* silence compile warning in FG-Matrix.cc due to signed vs unigned compare

+++ trunk/lib/FG-Matrix.cc
@@ -10,7 +10,7 @@ FGMatrix::FGMatrix(Image& image, unsigned int fg_threshold)
   for (; i!=end ; ++i) {
     data[row][line]=((*i).getL() < fg_threshold);
 
-    if (++row == image.w) {
+    if (++row == (unsigned int)image.w) {
       line++;
       row=0;
     }
